web/requirements.txt
Django
Pillow
numpy
django-ckeditor
django-widget-tweaks